Block-in-inline shouldn't keep an IFC from being non-contiguous.
A block-in-inline may only contain fragmented parallel flows, for all we
know, so it cannot unconditionally count as "in-flow", since we might
fail to mark a container as potentially non-contiguous that way.

In the test included, there's a text node "x xx", where the first word
should be in the first column, and the second word should be in the
fourth. There's no room for such a wide piece of text ("xx") before
we're past the tall float.

Bug: 346876226
Change-Id: I5bcf8840a1c71783dfa0bdce442a92eb5ddbf7b4